refactor: host-center 프로필 영역 컴포넌트로 분리, TempMain -> HostCenterPage 변경
